requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
barbar==0.2.1
blinker==1.4
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.28
fonttools==4.25.0
fsspec==2022.3.0
future==0.18.2
google-auth==2.6.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
joblib==0.17.0
kiwisolver==1.3.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
numexpr==2.8.1
numpy==1.22.3
oauthlib==3.2.0
openslide-python==1.1.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pip==22.0.4
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pyDeprecate==0.3.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytz==2021.3
pyu2f==0.1.5
PyYAML==5.4.1
reindent==3.5.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
Theano==1.0.5
threadpoolctl==2.1.0
torch==1.11.0
torchmetrics==0.7.3
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wheel==0.37.1
yarl==1.6.3
zipp==3.8.0
